pyproject.toml
[project]
name = "documentation-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.1.0",
    "certifi>=2026.1.4",
    "isort>=7.0.0",
    "langchain>=1.2.9",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.1",
    "langchain-openai>=1.1.7",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.8",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
